requirements.txt
absl-py==1.4.0
astunparse==1.6.3
av==10.0.0
backcall==0.2.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
easydict==1.9
english @ file:///home/conda/feedstock_root/build_artifacts/english_1620640779217/work
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.5.3
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1669279893622/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1676637751542/work
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
interchange @ file:///home/conda/feedstock_root/build_artifacts/interchange_1634204805395/work
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
neobolt @ file:///D:/bld/neobolt_1655667661601/work
neotime==1.7.4
numpy @ file:///D:/bld/numpy_1649806536290/work
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.1.5
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.20.3
psutil==5.9.4
py2neo @ file:///home/conda/feedstock_root/build_artifacts/py2neo_1634208170096/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyparsing==3.0.9
PySocks @ file:///D:/bld/pysocks_1648857448477/work
python-dateutil==2.8.2
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.22.1
scipy==1.7.3
seaborn==0.12.2
sentry-sdk==1.17.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1611376390914/work
SQLAlchemy==2.0.6
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.64.1
traitlets==5.9.0
typed-ast==1.5.4
typing_extensions==4.5.0
ultralytics==8.0.54
urllib3==1.26.14
utils==1.0.1
wcwidth==0.2.6
Werkzeug==2.2.3
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
wrapt==1.15.0
zipp==3.15.0
